+++ src/npm.ts
@@ -17,7 +17,7 @@ import { getConfig } from './config'
 import { logError } from './log'
 import { getNpmConfig } from './npmConfig'
 import { AsyncState, Dict, StrictDict } from './types'
-import { resolveCatalogVersion, resolveWorkspaceVersion } from './workspace'
+import { resolveCatalogVersion } from './workspace'
 
 export interface NpmLoader<T> {
   asyncstate: AsyncState
@@ -280,10 +280,6 @@ export const refreshPackageJsonData = (
 
     const promises = Object.entries(dependencies)
       .map(([dependencyName, version]) => {
-        if (version.startsWith('workspace:')) {
-          const resolved = resolveWorkspaceVersion(version, dependencyName, packageJsonFilePath)
-          return [dependencyName, resolved?.version ?? version] as const
-        }
         if (version.startsWith('catalog:')) {
           const resolved = resolveCatalogVersion(version, dependencyName, packageJsonFilePath)
           return [dependencyName, resolved?.version ?? version] as const